requirements.txt
mss
opencv-python
numpy
pytesseract
pyperclip
Pillow
keyboard